getting sparse adapter working
