+++ mttl/models/library/expert.py
@@ -164,7 +164,7 @@ def load_expert(
         os.path.join(expert_path, WEIGHTS_NAME)
     ):
         return load_expert_from_hf_checkpoint(expert_path, expert_name)
-    elif os.path.isdir(expert_path) and os.path.isfile(expert_path):
+    elif os.path.isdir(expert_path) or os.path.isfile(expert_path):
         return load_expert_from_pl_checkpoint(expert_path, expert_name)
 
     # this is not a local path, try to download from hub
@@ -193,7 +193,7 @@ def load_expert_from_pl_checkpoint(
 ):
     import os
 
-    from mttl.models.lightning.utils import download_from_hub
+    from mttl.models.lightning.base_module import download_from_hub
 
     if os.path.isfile(expert_path) or os.path.isdir(expert_path):
         expert_checkpoint = get_checkpoint_path(expert_path)
@@ -215,12 +215,6 @@ def load_expert_from_pl_checkpoint(
         # fix bug in checkpoints
         if "tokenizer" in expert_checkpoint["hyper_parameters"]:
             expert_checkpoint["hyper_parameters"].pop("tokenizer")
-
-        if not expert_info_data.get("expert_config", None):
-            expert_info_data["expert_config"] = expert_checkpoint["hyper_parameters"]
-        else:
-            if "tokenizer" in expert_info_data["expert_config"]:
-                expert_info_data["expert_config"].pop("tokenizer")
 
         if not expert_info_data.get("expert_name", None):
             expert_info_data["expert_name"] = expert_checkpoint["hyper_parameters"][
