+++ projects/sparse_finetuning/requirements.txt
@@ -1,0 +1,29 @@
+transformers==4.47.1 # need this specific version to avoid phi3 issues https://github.com/huggingface/transformers/issues/36071
+torch>=2.3.1
+datasets>=2.20.0
+pytorch-lightning>=2.3.3
+accelerate
+deepspeed
+huggingface_hub>=0.26.2
+click
+wandb
+rouge
+tqdm
+pandas
+sentence-transformers
+fsspec[adl]
+prettytable
+rich
+bitsandbytes
+matplotlib
+openai
+ray
+nevergrad
+evaluate
+seaborn
+azure-storage-blob
+azure-identity
+einops
+triton
+nltk
+# spops @ git+https://github.com/IST-DASLab/spops.git@main
